Add EventGenerator class
The EventGenerator is intended to be a way to turn a Poller into
an event generator. Instead of have a program continuously polling
the state of a device, this class does it for you and monitors
any changes, yielding tuples of button/axis states whenever the
state changes.

This class is iterable, so getting state changes is as simple as

    for event in EventGenerator(device):
        # do things with event

Note that since it's iterable, it's also possible to filter out
events with standard builtins like `filter` or generator
comprehensions.

It's an infinite data structure, so you need to avoid evaluating
the entire structure with a function like `list` or a list
comprehension. Stick with lazy methods that work on iterators.