Add FrameParser class
It is likely that we will need to combine the results of various
attributes of frames into single units. For example, joystick axes
are two units, but refer to the same joystick, and it might be
important to register a change in either axis as a change to the
entire joystick. This class does that.