Make resource imports portable
To deploy this project, we're building a .whl here so that you can install the
whole shebang with pip. What this means in practice is that you can't specify
filepaths of PNGs - we could be running smashbox_viewer.gui from anywhere. Not
only that, the package (with the PNGs inside it) could be anywhere - the global
Python installation, the user's own Python installation, or a venv.

As a result, there's a layer of abstraction that has to be done, and there's a
library importlib.resources that allows us to grab resources. There is now
a wrapper function that returns a context manager of these resources.

I also added Pillow to the list of dependencies, so attempting to install the
wheel will also install Pillow.