Add ambiguity handling to translator
Since it is possible for multiple button combinations to correspond
to the same controller state, the Translator needs to be able to
express ambiguity.

This adds a QuasiTranslator, which returns generators of possible
frames whenever polled. It also adds a RawTranslator, which
resolves this ambiguity by picking the first element of every
generator. More intelligent methods for resolving ambiguity will
likely be needed.